+++ app/page.tsx
@@ -1,6 +1,6 @@
 "use client" // TODO - this should be server component. Extract to individual client components
 
-import { useState } from "react"
+import { useRef, useState } from "react"
 import { Button } from "@/components/ui/button"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { ChevronRightIcon, ChevronLeftIcon } from "lucide-react"
@@ -50,7 +50,8 @@ const Home = () => {
   const [date, setDate] = useState(new Date())
   // 'id' holds the id of the category for which the task is being added
   const [newTaskInput, setNewTaskInput] = useState({ id: "", description: "" })
-  const [isDialogOpen, setIsDialogOpen] = useState(false)
+  const [isNewDialogOpen, setIsNewDialogOpen] = useState(false)
+  const [isEditDialogOpen, setIsEditDialogOpen] = useState(false)
 
   const [planner, setPlanner] = useState<Category[]>([
     {
@@ -114,6 +115,10 @@ const Home = () => {
     },
   ])
 
+  // These refs are used for managing the single-click and double-click functionalities.
+  const timerRef = useRef<NodeJS.Timeout | null>(null)
+  const doubleRef = useRef(false)
+
   const changeTaskStatus = (categoryId: string, taskId: string) => {
     const newPlanner = [...planner].map((category) => {
       if (category.id === categoryId) {
@@ -163,7 +168,7 @@ const Home = () => {
         return category
       })
       setPlanner(newPlanner)
-      setIsDialogOpen(false)
+      setIsNewDialogOpen(false)
     }
   }
 
@@ -192,7 +197,7 @@ const Home = () => {
           <ChevronRightIcon />
         </Button>
       </div>
-      <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
+      <Dialog open={isNewDialogOpen} onOpenChange={setIsNewDialogOpen}>
         <DialogTrigger>Add Task</DialogTrigger>
         <DialogContent>
           <DialogHeader>
@@ -204,7 +209,7 @@ const Home = () => {
                 items={categories}
                 value={newTaskInput.id}
                 onValueChange={(e) =>
-                  setNewTaskInput({ ...newTaskInput, id: e })
+                  setNewTaskInput({ ...newTaskInput, id: e! })
                 }
               >
                 <SelectTrigger className="w-[180px]">
@@ -259,9 +264,33 @@ const Home = () => {
                           ? "bg-red-900"
                           : ""
                   }
-                  // onClick={() => changeTaskStatus(category.id, task.id)}
-                  onDoubleClick={() => console.log("OKAY")}
-                  onLon
+                  onClick={() => {
+                    // The card can be single-clicked to change its status, or double-clicked to be edited.
+                    // To allow for double-clicking, a short timer is set when clicked once.
+
+                    // This timer is assigned to timerRef, to ensure that another timer cannot be activated when one is active.
+                    // Essentially, only one timer can be activated at a time.
+
+                    // The doubleRef ref is used to keep track of if the card has been double-clicked. Essentially, when a card
+                    // is clicked when there is no timer active, then the timer starts. If it is then clicked while a timer is active,
+                    // that will count as a double-click, so doubleRef is set to true. If the timer stops when doubleRef is false (when
+                    // there is no second click), then a single-click is registered.
+
+                    // doubleRef uses useRef instead of useState, since we must be able to change the state while the timer is active. useState
+                    // does not allow for this.
+                    if (timerRef.current === null) {
+                      timerRef.current = setTimeout(() => {
+                        timerRef.current = null
+                        if (doubleRef.current === false) {
+                          changeTaskStatus(category.id, task.id)
+                        }
+                        doubleRef.current = false
+                      }, 170)
+                    } else {
+                      setIsEditDialogOpen(true)
+                      doubleRef.current = true
+                    }
+                  }}
                 >
                   <CardContent>
                     <p>{task.description}</p>
@@ -272,52 +301,31 @@ const Home = () => {
           </div>
         ))}
       </div>
-
-      {/* <div className="flex w-full items-center justify-center gap-5">
-        <Button
-          variant="outline"
-          size="icon-xs"
-          onClick={() => {
-            date.setDate(date.getDate() - 1)
-            setDate(new Date(date))
-          }}
-        >
-          <ChevronLeftIcon />
-        </Button>
-        <h1>{date.toLocaleDateString()}</h1>
-        <Button
-          variant="outline"
-          size="icon-xs"
-          onClick={() => {
-            date.setDate(date.getDate() + 1)
-            setDate(new Date(date))
-          }}
-        >
-          <ChevronRightIcon />
-        </Button>
-      </div>
-      <Tabs defaultValue="Project" className="w-full">
-        <div className="flex w-full justify-center overflow-x-auto">
-          <TabsList className="shrink-0">
-            {planner.map((category) => (
-              <TabsTrigger key={category.id} value={category.name}>
-                {category.name}
-              </TabsTrigger>
-            ))}
-          </TabsList>
-        </div>
-        {planner.map((category) => (
-          <TabsContent key={category.id} value={category.name}>
-            <div className="flex flex-col items-stretch gap-3">
-              {category.tasks.map((task) => (
-                <div key={task.id}>
-                  <p>{task.description}</p>
-                </div>
-              ))}
-            </div>
-          </TabsContent>
-        ))}
-      </Tabs> */}
+      <Dialog open={isEditDialogOpen} onOpenChange={setIsEditDialogOpen}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Edit Task</DialogTitle>
+          </DialogHeader>
+          <FieldGroup>
+            {/* <Field>
+              <Label htmlFor="description">Description</Label>
+              <Input
+                id="description"
+                name="description"
+                value={newTaskInput.description}
+                onChange={(e) =>
+                  setNewTaskInput({
+                    ...newTaskInput,
+                    description: e.target.value,
+                  })
+                }
+              />
+            </Field> */}
+            <Button>Update</Button>
+            <Button>Cancel</Button>
+          </FieldGroup>
+        </DialogContent>
+      </Dialog>
     </div>
   )
 }